Combine VM with Statics, and rely on JIT inlining of releaseFence
