+++ src/library/scala/runtime/ScalaRunTime.scala
@@ -268,9 +268,4 @@ object ScalaRunTime {
   def wrapShortArray(xs: Array[Short]): ArraySeq[Short] = if (xs ne null) new ArraySeq.ofShort(xs) else null
   def wrapBooleanArray(xs: Array[Boolean]): ArraySeq[Boolean] = if (xs ne null) new ArraySeq.ofBoolean(xs) else null
   def wrapUnitArray(xs: Array[Unit]): ArraySeq[Unit] = if (xs ne null) new ArraySeq.ofUnit(xs) else null
-
-  /**
-    * Forwards to `VarHandle.releaseFence` or `sun.misc.Unsafe.storeFence` on JDK9+ or JDK8 respectively.
-    */
-  @inline def releaseFence(): Unit = VM.RELEASE_FENCE.invoke()
 }
