+++ src/library/scala/runtime/Statics.java
@@ -1,5 +1,10 @@
 package scala.runtime;
 
+import java.lang.invoke.MethodHandle;
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.MethodType;
+import java.lang.reflect.Field;
+
 /** Not for public consumption.  Usage by the runtime only.
  */
 
@@ -117,4 +122,48 @@ public final class Statics {
 
   /** Used as a marker object to return from PartialFunctions */
   public static final Object pfMarker = new Object();
+
+  // @ForceInline would be nice here.
+  public static void releaseFence() throws Throwable {
+    VM.RELEASE_FENCE.invoke();
+  }
+
+  final static class VM {
+      static final MethodHandle RELEASE_FENCE;
+
+      static {
+          RELEASE_FENCE = mkHandle();
+      }
+
+      private static MethodHandle mkHandle() {
+          MethodHandles.Lookup lookup = MethodHandles.lookup();
+          try {
+              return lookup.findStatic(Class.forName("java.lang.invoke.VarHandle"), "releaseFence", MethodType.methodType(Void.TYPE));
+          } catch (ClassNotFoundException e) {
+              try {
+                  Class<?> unsafeClass = Class.forName("sun.misc.Unsafe");
+                  return lookup.findVirtual(unsafeClass, "storeFence", MethodType.methodType(void.class)).bindTo(findUnsafe(unsafeClass));
+              } catch (NoSuchMethodException | ClassNotFoundException | IllegalAccessException e1) {
+                  ExceptionInInitializerError error = new ExceptionInInitializerError(e1);
+                  error.addSuppressed(e);
+                  throw error;
+              }
+          } catch (NoSuchMethodException | IllegalAccessException e) {
+              throw new ExceptionInInitializerError(e);
+          }
+      }
+
+      private static Object findUnsafe(Class<?> unsafeClass) throws IllegalAccessException {
+          Object found = null;
+          for (Field field : unsafeClass.getDeclaredFields()) {
+              if (field.getType() == unsafeClass) {
+                  field.setAccessible(true);
+                  found = field.get(null);
+                  break;
+              }
+          }
+          if (found == null) throw new IllegalStateException("No instance of Unsafe found");
+          return found;
+      }
+  }
 }
